refactor: extract handlers into _handlers.py and create platform/ subpackage
Extract _handle_invoke, _handle_stream, _handle_state logic from
LangGraphApp into standalone functions in _handlers.py. LangGraphApp
retains thin delegation stubs to preserve backward-compatible test
call patterns. _error_response moves with the handlers.

Create azure_functions_langgraph/platform/ subpackage with __init__.py
for upcoming LangGraph Platform API compatibility layer (v0.3.0).

- app.py: 436 → 297 lines (orchestration + route wiring only)
- _handlers.py: 215 lines (all request handling logic)
- platform/__init__.py: docstring placeholder for #36-#39
- Zero test changes — all 105 tests pass
- Coverage: 98.50% (up from 98.43%)
- Lint + typecheck: clean

Closes #35
